Rename WrappedShape -> WrappedDrawing
